refactor: tests - remove unused functions, imports etc.
