+++ healthcare/healthcare/doctype/patient_history_settings/test_patient_history_settings.py
@@ -7,9 +7,6 @@ import json
 import frappe
 from frappe.utils import getdate, strip_html
 
-from healthcare.healthcare.doctype.patient_appointment.test_patient_appointment import (
-	create_patient,
-)
 from healthcare.tests.utils import HealthcareTestSuite
