+++ healthcare/tests/utils.py
@@ -603,6 +603,16 @@ class BootStrapTestData:
 				"first_name": "_Test IPD Patient",
 				"sex": "Female",
 			},
+			{
+				"doctype": "Patient",
+				"first_name": "_Test Patient 2",
+				"sex": "Female",
+			},
+			{
+				"doctype": "Patient",
+				"first_name": "_Test Patient 3",
+				"sex": "Female",
+			},
 		]
 		self.make_records(["first_name"], records)
